Delete Publisher

+++ src/main/java/com/higgsup/intern/ebshop/service/impl/PublisherService.java
@@ -3,7 +3,6 @@ package com.higgsup.intern.ebshop.service.impl;
 import com.higgsup.intern.ebshop.dto.PublisherDTO;
 import com.higgsup.intern.ebshop.exception.ServiceException;
 import com.higgsup.intern.ebshop.jdbc.dao.PublisherDAO;
-import com.higgsup.intern.ebshop.jdbc.model.Publisher;
 import com.higgsup.intern.ebshop.service.IPublisherService;
 import ma.glasnost.orika.MapperFacade;
 import org.springframework.stereotype.Service;
@@ -21,6 +20,11 @@ public class PublisherService implements IPublisherService {
 
     @Override
     public void delete(Long id) {
+
+        if(publisherDAO.countBook(id) > 0)
+        {
+            throw new ServiceException(String.format("Publisher with id = %d still have book in it!", id));
+        }
         publisherDAO.delete(id);
     }
 }
